Fix some bugs, update executable, update README.md
Fixed some issues with the menu, allowed for more viewable displays of
the array's content. Fixed some bugs relating to the insertion of
elements at specific indices and increasing the size of the array
dynamically. Also changed pop() so that the the capacity of the
array is reduced by half when the size is 1/4 of the capacity.